backend/requirements.txt
asgiref==3.9.1
attrs==25.3.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
defusedxml==0.7.1
Django==5.0
django-cors-headers==3.13.0
djangorestframework==3.15.0
djangorestframework_simplejwt==5.5.1
djoser==2.3.1
flake8==6.0.0
gunicorn==23.0.0
idna==3.10
iniconfig==2.1.0
mccabe==0.7.0
oauthlib==3.3.1
packaging==25.0
pillow==11.2.1
pluggy==0.13.1
psycopg2-binary==2.9.10
py==1.11.0
pycodestyle==2.10.0
pycparser==2.22
pyflakes==3.0.1
PyJWT==2.10.1
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python3-openid==3.2.0
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.7.0
sqlparse==0.5.3
toml==0.10.2
urllib3==2.5.0
webcolors==1.11.1
